Refactored background property into a shared resource.
