+++ src/components/card/card.scss
@@ -66,39 +66,13 @@
     margin-top: 2rem;
   }
 }
-
+/* Card elements
+ * ====================== */
 .card__media {
   margin-left: -2rem;
   margin-right: -2rem;
   margin-top: -2rem;
   margin-bottom: 2rem;
-  text-align: center;
-}
-
-.card--borderless {
-  border-width: 0;
-  padding: 0;
-
-  .card__media {
-    margin-left: 0;
-    margin-right: 0;
-    margin-top: 0;
-  }
-
-  .bttn {
-    margin-bottom: 0;
-  }
-}
-
-.card--media-padded {
-  .card__media {
-    margin-left: 0;
-    margin-top: 0;
-    margin-right: 0;
-  }
-}
-
-.card--centered {
   text-align: center;
 }
 
@@ -183,6 +157,35 @@
   .svg-inline--fa {
     width: 17px;
   }
+}
+
+/* Card modifiers
+ * ====================== */
+.card--borderless {
+  border-width: 0;
+  padding: 0;
+
+  .card__media {
+    margin-left: 0;
+    margin-right: 0;
+    margin-top: 0;
+  }
+
+  .bttn {
+    margin-bottom: 0;
+  }
+}
+
+.card--media-padded {
+  .card__media {
+    margin-left: 0;
+    margin-top: 0;
+    margin-right: 0;
+  }
+}
+
+.card--centered {
+  text-align: center;
 }
 
 /* Card media styles
