+++ src/components/utlity.ts
@@ -1,0 +1,3 @@
+export function className(text: string): string {
+  return text.replace(/_/g, '-');
+}
